+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-20T04:24:04.882Z")
+    created: Optional[datetime] = parse_datetime("2026-01-20T11:19:15.524Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")