+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfctasr7ya94j37xw1rg3dfd"
+        "conv_01kfdj30dadcb7tb6s6vhxzcbw"
     )
     r"""Unique conversation identifier with `conv_` prefix."""