+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-20T04:24:01.381Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-20T11:19:12.098Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")