+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFCTATK71EK89P7V7XTXVV48"
+        "01KFDJ30X33MMRJ0VB0YYC0CP2"
     )
     r"""The unique identifier of the data source"""