+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfctasrgh2kaq8thvt3d5mjg"
+        "conv_01kfdj30dfaz93dap315azgqxz"
     )
     r"""Unique conversation identifier with `conv_` prefix."""