+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFCTATKBFM4FFD7SE4QGR4TJ"
+        "01KFDJ30XGYWV9MENTKFACFQW4"
     )
     r"""The unique identifier of the data source"""