+++ packages/orq-rc/RELEASES.md
@@ -4839,3 +4839,13 @@ Based on:
 - [python v4.2.0-rc.46] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.46] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.46 - packages/orq-rc
+
+## 2026-01-20 11:21:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.690.0 (2.797.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.47] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.47 - packages/orq-rc